requirements/base.txt
Django==4.2
cytoolz==0.12.1
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg==3.1.8
